backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
openai
git+https://github.com/cisco-ai-defense/a2a-scanner.git
